@@ -2,7 +2,7 @@ package bucket
 
 import (
 	"github.com/neo9/mongodb-backups/pkg/config"
-	log "github.com/sirupsen/logrus"
+	"github.com/neo9/mongodb-backups/pkg/log"
 )
 
 type Bucket interface {
@@ -14,13 +14,13 @@ type Bucket interface {
 
 func New(bucket *config.Bucket) Bucket {
 	if bucket.S3.Name != "" {
-		log.Infof("using s3 storage (%s)", bucket.S3.Name)
+		log.Info("using s3 storage (%s)", bucket.S3.Name)
 		return NewS3Bucket(&bucket.S3)
 	} else if bucket.GS.Name != "" {
-		log.Infof("using gs storage (%s)", bucket.GS.Name)
+		log.Info("using gs storage (%s)", bucket.GS.Name)
 		return NewGSBucket(&bucket.GS)
 	} else if bucket.Minio.Name != "" {
-		log.Infof("using Minio storage (%s)", bucket.Minio.Name)
+		log.Info("using Minio storage (%s)", bucket.Minio.Name)
 		return NewMinioBucket(&bucket.Minio)
 	}
 
