@@ -5,10 +5,9 @@ import (
 	"path"
 	"time"
 
-	"github.com/codeskyblue/go-sh"
 	"github.com/neo9/mongodb-backups/pkg/config"
+	"github.com/neo9/mongodb-backups/pkg/log"
 	"github.com/neo9/mongodb-backups/pkg/utils"
-	"github.com/prometheus/common/log"
 )
 
 type MongoDBDump struct {
@@ -18,23 +17,23 @@ type MongoDBDump struct {
 }
 
 func CreateDump(plan *config.Plan) (MongoDBDump, error) {
-    var err error
-    var mongoDBDump MongoDBDump
+	var err error
+	var mongoDBDump MongoDBDump
 
-    maxRetries := plan.CreateDump.MaxRetries
-    retryDelay := plan.CreateDump.RetryDelay * time.Second
+	maxRetries := plan.CreateDump.MaxRetries
+	retryDelay := plan.CreateDump.RetryDelay * time.Second
 
-    for i := 0; i < maxRetries; i++ {
-        mongoDBDump, err = CreateDumpInternal(plan)
-        if err != nil {
-            log.Errorf("Error creating mongodump (retry %d/%d): %v", i+1, maxRetries, err)
-            time.Sleep(retryDelay)
-        } else {
-            return mongoDBDump, nil
-        }
-    }
+	for i := 0; i < maxRetries; i++ {
+		mongoDBDump, err = CreateDumpInternal(plan)
+		if err != nil {
+			log.Error("Error creating mongodump (retry %d/%d): %v", i+1, maxRetries, err)
+			time.Sleep(retryDelay)
+		} else {
+			return mongoDBDump, nil
+		}
+	}
 
-    return mongoDBDump, err
+	return mongoDBDump, err
 }
 
 func CreateDumpInternal(plan *config.Plan) (MongoDBDump, error) {
@@ -55,24 +54,22 @@ func CreateDumpInternal(plan *config.Plan) (MongoDBDump, error) {
 
 	duration, err := utils.GetDurationFromTimeString(plan.Timeout)
 	if err != nil {
-		log.Errorf("Error parsing timeout: %v", err)
+		log.Error("Error parsing timeout: %v", err)
 		return mongoDBDump, err
 	}
 
 	startTime := time.Now()
-	output, err := sh.Command("/bin/sh", "-c", dumpCommand).
-		SetTimeout(duration).
-		CombinedOutput()
+	output, err := utils.LaunchCommand(dumpCommand, duration)
 	mongoDBDump.Duration = time.Since(startTime).Seconds()
 
 	if err != nil {
-		log.Errorf("Error creating dump: %v, %s", err, output)
-		log.Errorf("Dump timeout: %s", duration)
+		log.Error("Error creating dump: %v, %s", err, output)
+		log.Error("Dump timeout: %s", duration)
 		RemoveFile(mongoDBDump.ArchiveFile)
 		return mongoDBDump, err
 	}
 
-	log.Infof("Done creating dump for %s", plan.Name)
+	log.Info("Done creating dump for %s", plan.Name)
 	err = logToFile(mongoDBDump.LogFile, output)
 	if err != nil {
 		RemoveFile(mongoDBDump.ArchiveFile)
