@@ -4,10 +4,9 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/codeskyblue/go-sh"
 	"github.com/neo9/mongodb-backups/pkg/config"
+	"github.com/neo9/mongodb-backups/pkg/log"
 	"github.com/neo9/mongodb-backups/pkg/utils"
-	"github.com/prometheus/common/log"
 )
 
 func RestoreDump(filename string, args string, plan *config.Plan) error {
@@ -23,23 +22,21 @@ func RestoreDump(filename string, args string, plan *config.Plan) error {
 
 	duration, err := utils.GetDurationFromTimeString(plan.Timeout)
 	if err != nil {
-		log.Errorf("Error parsing timeout: %v", err)
+		log.Error("Error parsing timeout: %v", err)
 		return err
 	}
 
 	startTime := time.Now()
-	output, err := sh.Command("/bin/sh", "-c", restoreCommand).
-		SetTimeout(duration).
-		CombinedOutput()
+	output, err := utils.LaunchCommand(restoreCommand, duration)
 	seconds := time.Since(startTime).Seconds()
 
 	if err != nil {
-		log.Errorf("Error restoring dump: %v, %s", err, output)
-		log.Errorf("Restoring timeout: %s", duration)
+		log.Error("Error restoring dump: %v, %s", err, output)
+		log.Error("Restoring timeout: %s", duration)
 		return err
 	}
 
-	log.Infof("Done restoring dump for %s in %s s", plan.Name, seconds)
+	log.Info("Done restoring dump for %s in %f s", plan.Name, seconds)
 
 	return err
 }
