@@ -11,12 +11,12 @@ func GetBucketFileTimestamp(file string) (int64, error) {
 	reg := regexp.MustCompile(`mongodb-snapshot-(?P<Time>\d+)\.(gz|log)`)
 	match := reg.FindStringSubmatch(file)
 	if len(match) != 3 {
-		return 0, fmt.Errorf("File does not match pattern in folder: " + file)
+		return 0, fmt.Errorf("file does not match pattern in folder: %s", file)
 	}
 
 	timestamp, err := strconv.ParseInt(match[1], 10, 64)
 	if err != nil {
-		return 0, fmt.Errorf("File has invalid timestamp in folder: " + file)
+		return 0, fmt.Errorf("file has invalid timestamp in folder: %s ", file)
 	}
 
 	return timestamp, nil
